Only change the line height of the view line
Previously, if a decoration specified a custom line height, this
affected all viewlines that are covered by model line. Now the line
height is only applied to the view lines that render the decoration.

It is still possible to change the line height of an entire view line by
setting the decoration option `isWholeLine`.