Fix out of bounds text selection with line wrapping
The original behaviour is confusing and somewhat hard to explain. Given
a Monaco editor with text wrapping enabled, if the user starts selecting
text and moves the mouse outside of the editor viewbox, to the right of
the screen, the cursor in the editor would be somewhere halfway the
editor. The editor cursor did move with the mouse. So moving the mouse
further right, moved the cursor further forward. The text cursor could
even move onto the next line, below the mouse position. In case of RTL,
the behaviour was the same. Further right means more forward, meaning it
was visually mirrored.

The new behaviour with this PR is easier to explain. If the cursor moves
out of bounds while selecting text, the selection will match the end of
the wrapped line at column matching the cursor height. It also takes RTL
into account.

The new behaviour matches gedit, LibreOffice, Google Docs, HTML content
in browsers, and even Monaco Editor with text wrapping disabled.